requirements.txt
streamlit
tiktoken
python-dotenv
requests
